fix task_3

+++ week_1/task_3.cpp
@@ -7,7 +7,7 @@ int main() {
     double a, b, c;
     cin >> a >> b >> c;
     
-    int D = b * b - 4 * a * c;
+    double D = b * b - 4 * a * c;
     if (D == 0) {
         cout << -b / 2 / a;
     } else if (D > 0) {
